+++ esphome/components/ds1307/ds1307.cpp
@@ -37,14 +37,14 @@ void DS1307Component::read_time() {
     ESP_LOGW(TAG, "RTC halted, not syncing to system clock.");
     return;
   }
-  time::ESPTime rtc_time{.second = uint8_t(ds1307_.reg.second + 10 * ds1307_.reg.second_10),
-                         .minute = uint8_t(ds1307_.reg.minute + 10u * ds1307_.reg.minute_10),
-                         .hour = uint8_t(ds1307_.reg.hour + 10u * ds1307_.reg.hour_10),
-                         .day_of_week = uint8_t(ds1307_.reg.weekday),
-                         .day_of_month = uint8_t(ds1307_.reg.day + 10u * ds1307_.reg.day_10),
-                         .day_of_year = 1,  // ignored by recalc_timestamp_utc(false)
-                         .month = uint8_t(ds1307_.reg.month + 10u * ds1307_.reg.month_10),
-                         .year = uint16_t(ds1307_.reg.year + 10u * ds1307_.reg.year_10 + 2000)};
+  ESPTime rtc_time{.second = uint8_t(ds1307_.reg.second + 10 * ds1307_.reg.second_10),
+                   .minute = uint8_t(ds1307_.reg.minute + 10u * ds1307_.reg.minute_10),
+                   .hour = uint8_t(ds1307_.reg.hour + 10u * ds1307_.reg.hour_10),
+                   .day_of_week = uint8_t(ds1307_.reg.weekday),
+                   .day_of_month = uint8_t(ds1307_.reg.day + 10u * ds1307_.reg.day_10),
+                   .day_of_year = 1,  // ignored by recalc_timestamp_utc(false)
+                   .month = uint8_t(ds1307_.reg.month + 10u * ds1307_.reg.month_10),
+                   .year = uint16_t(ds1307_.reg.year + 10u * ds1307_.reg.year_10 + 2000)};
   rtc_time.recalc_timestamp_utc(false);
   if (!rtc_time.is_valid()) {
     ESP_LOGE(TAG, "Invalid RTC time, not syncing to system clock.");
