+++ esphome/components/lcd_base/lcd_display.h
@@ -1,11 +1,7 @@
 #pragma once
 
 #include "esphome/core/component.h"
-#include "esphome/core/defines.h"
-
-#ifdef USE_TIME
-#include "esphome/components/time/real_time_clock.h"
-#endif
+#include "esphome/core/time.h"
 
 #include <map>
 #include <vector>
@@ -44,13 +40,10 @@ class LCDDisplay : public PollingComponent {
   /// Evaluate the printf-format and print the text at column=0 and row=0.
   void printf(const char *format, ...) __attribute__((format(printf, 2, 3)));
 
-#ifdef USE_TIME
   /// Evaluate the strftime-format and print the text at the specified column and row.
-  void strftime(uint8_t column, uint8_t row, const char *format, time::ESPTime time)
-      __attribute__((format(strftime, 4, 0)));
+  void strftime(uint8_t column, uint8_t row, const char *format, ESPTime time) __attribute__((format(strftime, 4, 0)));
   /// Evaluate the strftime-format and print the text at column=0 and row=0.
-  void strftime(const char *format, time::ESPTime time) __attribute__((format(strftime, 2, 0)));
-#endif
+  void strftime(const char *format, ESPTime time) __attribute__((format(strftime, 2, 0)));
 
   /// Load custom char to given location
   void loadchar(uint8_t location, uint8_t charmap[]);
