+++ esphome/components/max7219digit/max7219digit.cpp
@@ -325,18 +325,16 @@ uint8_t MAX7219Component::printdigitf(const char *format, ...) {
   return 0;
 }
 
-#ifdef USE_TIME
-uint8_t MAX7219Component::strftimedigit(uint8_t pos, const char *format, time::ESPTime time) {
+uint8_t MAX7219Component::strftimedigit(uint8_t pos, const char *format, ESPTime time) {
   char buffer[64];
   size_t ret = time.strftime(buffer, sizeof(buffer), format);
   if (ret > 0)
     return this->printdigit(pos, buffer);
   return 0;
 }
-uint8_t MAX7219Component::strftimedigit(const char *format, time::ESPTime time) {
+uint8_t MAX7219Component::strftimedigit(const char *format, ESPTime time) {
   return this->strftimedigit(0, format, time);
 }
-#endif
 
 }  // namespace max7219digit
 }  // namespace esphome
