+++ esphome/components/sun/sun.h
@@ -1,8 +1,10 @@
 #pragma once
 
+#include "esphome/core/automation.h"
 #include "esphome/core/component.h"
 #include "esphome/core/helpers.h"
-#include "esphome/core/automation.h"
+#include "esphome/core/time.h"
+
 #include "esphome/components/time/real_time_clock.h"
 
 namespace esphome {
@@ -26,7 +28,7 @@ struct GeoLocation {
 };
 
 struct Moment {
-  time::ESPTime dt;
+  ESPTime dt;
 
   num_t jd() const;
   num_t jde() const;
@@ -57,18 +59,18 @@ class Sun {
   void set_latitude(double latitude) { location_.latitude = latitude; }
   void set_longitude(double longitude) { location_.longitude = longitude; }
 
-  optional<time::ESPTime> sunrise(double elevation);
-  optional<time::ESPTime> sunset(double elevation);
-  optional<time::ESPTime> sunrise(time::ESPTime date, double elevation);
-  optional<time::ESPTime> sunset(time::ESPTime date, double elevation);
+  optional<ESPTime> sunrise(double elevation);
+  optional<ESPTime> sunset(double elevation);
+  optional<ESPTime> sunrise(ESPTime date, double elevation);
+  optional<ESPTime> sunset(ESPTime date, double elevation);
 
   double elevation();
   double azimuth();
 
  protected:
   internal::HorizontalCoordinate calc_coords_();
-  optional<time::ESPTime> calc_event_(bool rising, double zenith);
-  optional<time::ESPTime> calc_event_(time::ESPTime date, bool rising, double zenith);
+  optional<ESPTime> calc_event_(bool rising, double zenith);
+  optional<ESPTime> calc_event_(ESPTime date, bool rising, double zenith);
 
   time::RealTimeClock *time_;
   internal::GeoLocation location_;
