+++ esphome/components/sun/text_sensor/sun_text_sensor.h
@@ -1,6 +1,8 @@
 #pragma once
 
 #include "esphome/core/component.h"
+#include "esphome/core/time.h"
+
 #include "esphome/components/sun/sun.h"
 #include "esphome/components/text_sensor/text_sensor.h"
 
@@ -15,7 +17,7 @@ class SunTextSensor : public text_sensor::TextSensor, public PollingComponent {
   void set_format(const std::string &format) { format_ = format; }
 
   void update() override {
-    optional<time::ESPTime> res;
+    optional<ESPTime> res;
     if (this->sunrise_) {
       res = this->parent_->sunrise(this->elevation_);
     } else {
