+++ esphome/components/time/automation.cpp
@@ -1,5 +1,7 @@
 #include "automation.h"
+
 #include "esphome/core/log.h"
+
 #include <cinttypes>
 
 namespace esphome {
