+++ esphome/components/tm1637/tm1637.cpp
@@ -368,16 +368,14 @@ uint8_t TM1637Display::printf(const char *format, ...) {
   return 0;
 }
 
-#ifdef USE_TIME
-uint8_t TM1637Display::strftime(uint8_t pos, const char *format, time::ESPTime time) {
+uint8_t TM1637Display::strftime(uint8_t pos, const char *format, ESPTime time) {
   char buffer[64];
   size_t ret = time.strftime(buffer, sizeof(buffer), format);
   if (ret > 0)
     return this->print(pos, buffer);
   return 0;
 }
-uint8_t TM1637Display::strftime(const char *format, time::ESPTime time) { return this->strftime(0, format, time); }
-#endif
+uint8_t TM1637Display::strftime(const char *format, ESPTime time) { return this->strftime(0, format, time); }
 
 }  // namespace tm1637
 }  // namespace esphome
