+++ esphome/components/tm1638/tm1638.cpp
@@ -211,16 +211,14 @@ uint8_t TM1638Component::printf(const char *format, ...) {
   return 0;
 }
 
-#ifdef USE_TIME
-uint8_t TM1638Component::strftime(uint8_t pos, const char *format, time::ESPTime time) {
+uint8_t TM1638Component::strftime(uint8_t pos, const char *format, ESPTime time) {
   char buffer[64];
   size_t ret = time.strftime(buffer, sizeof(buffer), format);
   if (ret > 0)
     return this->print(pos, buffer);
   return 0;
 }
-uint8_t TM1638Component::strftime(const char *format, time::ESPTime time) { return this->strftime(0, format, time); }
-#endif
+uint8_t TM1638Component::strftime(const char *format, ESPTime time) { return this->strftime(0, format, time); }
 
 //////////////// SPI   ////////////////
