+++ esphome/components/tuya/tuya.cpp
@@ -508,7 +508,7 @@ void Tuya::send_wifi_status_() {
 void Tuya::send_local_time_() {
   std::vector<uint8_t> payload;
   auto *time_id = *this->time_id_;
-  time::ESPTime now = time_id->now();
+  ESPTime now = time_id->now();
   if (now.is_valid()) {
     uint8_t year = now.year - 2000;
     uint8_t month = now.month;
