+++ esphome/components/tuya/tuya.h
@@ -7,6 +7,7 @@
 
 #ifdef USE_TIME
 #include "esphome/components/time/real_time_clock.h"
+#include "esphome/core/time.h"
 #endif
 
 #include <vector>
